Fix detection of errors around highest pallets when zero tokens
There was an assumption previously that tokens not present in a
district would not be included in the mapping, rather than being
included with a zero count. This was incorrect due to how the
scorer UI works and not neccesary as we're using a Counter type
for this collection. This commit fixes this by handling explicit
zeros.

Fixes https://github.com/srobo/sr2025-comp/issues/17